[#47] ♻️ Refactor: 메서드 이름 및 변수 타입 변경
- 주간 끄기 제한 횟수 상수로 변경

+++ src/main/java/akuma/whiplash/domains/alarm/application/dto/response/AlarmRemainingOffCountResponse.java
@@ -10,5 +10,5 @@ public record AlarmRemainingOffCountResponse(
         description = "회원의 남은 알람 끄기 횟수(회원당 매주 2회 부여, 매주 월요일 초기화)",
         example = "1"
     )
-    Long remainingOffCount
+    int remainingOffCount
 ) {}
